scaffold: spatio_flux/composites/ package + shared _constants extraction
