+++ spatio_flux/composites/spatial.py
@@ -1,0 +1,2 @@
+"""Composite generators — spatial group. See spec at
+docs/superpowers/specs/2026-05-12-composite-generator-convention.md."""
