+++ spatio_flux/experiments/test_suite.py
@@ -46,23 +46,14 @@ from spatio_flux.processes import (
 # Constants
 # ---------------------------------------------------------------------
 
-SQUARE_BOUNDS = (50.0, 50.0)  # this is in um
-SQUARE_BINS = (10, 10)
-
-DEFAULT_BOUNDS = (40.0, 80.0)
-DEFAULT_BINS = (10, 20)
-DEFAULT_BINS_SMALL = (2, 4)
-DEFAULT_ADVECTION = (0.0, 0.2)
-DEFAULT_DIFFUSION = 0.5
-DEFAULT_ADD_RATE = 0.1
-DEFAULT_ADD_BOUNDARY = ['top', 'left', 'right']
-DEFAULT_REMOVE_BOUNDARY = ['left', 'right']
-DEFAULT_INITIAL_MIN_MAX = {
-        'glucose': (10, 10),
-        'acetate': (0, 0),
-        'dissolved biomass': (0, 0.1),
-        # 'detritus': (0, 0)
-    }
+from spatio_flux.composites._constants import (  # noqa: F401
+    SQUARE_BOUNDS, SQUARE_BINS,
+    DEFAULT_BOUNDS, DEFAULT_BINS, DEFAULT_BINS_SMALL,
+    DEFAULT_ADVECTION, DEFAULT_DIFFUSION,
+    DEFAULT_ADD_RATE, DEFAULT_ADD_BOUNDARY, DEFAULT_REMOVE_BOUNDARY,
+    DEFAULT_INITIAL_MIN_MAX,
+    get_newtonian_particles_process,
+)
 
 DEFAULT_RUNTIME_SHORT = 10
 DEFAULT_RUNTIME_LONG = 60
@@ -690,19 +681,6 @@ def plot_particle_dfba_comets(results, state, config=None):
 
 
 # ---- PYMUNK PARTICLES ------------------------------------------------
-
-def get_newtonian_particles_process(config=None):
-    return {
-        '_type': 'process',
-        'address': 'local:PymunkParticleMovement',
-        'config': config,
-        'inputs': {
-            'particles': ['particles'],
-        },
-        'outputs': {
-            'particles': ['particles'],
-        }
-    }
 
 def get_newtonian_particles_doc(core=None, config=None):
     n_particles = config.get('n_particles', 1)
